requirements.txt
fastapi==0.79.0
uvicorn==0.18.2
emoji==1.7.0
requests==2.23.0
botbuilder-core==4.14.2
botframework-connector==4.14.2
botbuilder-schema==4.14.2
botbuilder-dialogs==4.14.2
botbuilder-ai==4.14.2
botbuilder-applicationinsights==4.14.2
botbuilder-integration-aiohttp==4.14.2
botbuilder-integration-applicationinsights-aiohttp==4.14.2
botbuilder-testing==4.14.2
datatypes-date-time==1.0.0.a2
azure-cognitiveservices-language-luis==0.2.0
msrest==0.6.19
aiohttp==3.7.4.post0
azure-identity==1.5.0
azure-keyvault==4.1.0
opencensus-ext-azure==1.1.6
opencensus-ext-logging==0.1.1
opencensus-ext-requests==0.8.0